Move calculatePremium to cover integration utils
